# logistic-package-api/swagger/logistic_package_api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "logistic_package_api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "LogisticPackageApiService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/packages": {
      "get": {
        "operationId": "LogisticPackageApiService_ListV1",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListV1Response"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "LogisticPackageApiService"
        ]
      },
      "post": {
        "operationId": "LogisticPackageApiService_CreateV1",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateResponseV1"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateRequestV1"
            }
          }
        ],
        "tags": [
          "LogisticPackageApiService"
        ]
      },
      "put": {
        "operationId": "LogisticPackageApiService_UpdateV1",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateV1Response"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1UpdateV1Request"
            }
          }
        ],
        "tags": [
          "LogisticPackageApiService"
        ]
      }
    },
    "/v1/packages/{packageId}": {
      "get": {
        "operationId": "LogisticPackageApiService_GetV1",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetV1Response"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "packageId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "LogisticPackageApiService"
        ]
      },
      "delete": {
        "operationId": "LogisticPackageApiService_DeleteV1",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteV1Response"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "packageId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "LogisticPackageApiService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1CreateRequestV1": {
      "type": "object",
      "properties": {
        "value": {
          "$ref": "#/definitions/v1Package"
        }
      }
    },
    "v1CreateResponseV1": {
      "type": "object",
      "properties": {
        "packageId": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "v1DeleteV1Response": {
      "type": "object"
    },
    "v1GetV1Response": {
      "type": "object",
      "properties": {
        "value": {
          "$ref": "#/definitions/v1Package"
        }
      }
    },
    "v1ListV1Response": {
      "type": "object",
      "properties": {
        "packages": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Package"
          }
        }
      }
    },
    "v1Package": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        },
        "title": {
          "type": "string"
        },
        "weight": {
          "type": "string",
          "format": "uint64"
        },
        "created": {
          "type": "string",
          "format": "date-time"
        },
        "updated": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "v1UpdateV1Request": {
      "type": "object",
      "properties": {
        "value": {
          "$ref": "#/definitions/v1Package"
        }
      }
    },
    "v1UpdateV1Response": {
      "type": "object"
    }
  }
}